final update

+++ Backend/Bridge/src/read_buffer.h
@@ -4,13 +4,13 @@
 
 class ReadBuffer {
     public:
-        ReadBuffer(uint8_t buffer_size);
+        ReadBuffer(uint16_t buffer_size);
         void clean();
         void append(const char input_byte);
         const char* read();
 
     private:
         char* read_buffer;
-        uint8_t read_buffer_size;
-        uint8_t read_buffer_index = 0;
+        uint16_t read_buffer_size;
+        uint16_t read_buffer_index = 0;
 };
